+++ src/components/App/App.styled.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const MainContainer = styled("main")`
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+`;
+
+export const TopSwitcherWrapper = styled("div")`
+  margin-bottom: 5px;
+`;
+
+export const BottomSwitcherWrapper = styled("div")`
+  margin-top: 5px;
+`;
+
+export const CalendarWrapper = styled("div")`
+  flex-grow: 1;
+  overflow: hidden;
+`;
