+++ src/components/Calendar/Calendar.tsx
@@ -1,36 +1,13 @@
-import styled, { css } from "styled-components";
 import { useSelector } from "react-redux";
 
 import Day from "../Day";
 import useCalendar from "@/hooks/useCalendar";
 
+import { RootState } from "@/redux/store";
 import { calculateMonthCalendar, calculateWeekCalendar } from "../../helpers/calculateCalendar";
 import { selectTasksByDates } from "@/redux/tasks/selectors";
 import { areDatesEqual } from "@/utils/date";
-import { RootState } from "@/redux/store";
-
-export const Wrapper = styled("div")`
-  height: 100%;
-  background-color: #ebeeed;
-`;
-
-export const ListWrapper = styled("ul")<{ $monthmode: boolean }>`
-  display: grid;
-  height: 100%;
-  grid-template-columns: repeat(7, minmax(0, 1fr));
-
-  ${p =>
-    p.$monthmode &&
-    css`
-      grid-template-rows: repeat(6, minmax(0, 1fr));
-    `}
-
-  gap: 5px;
-`;
-
-export const Item = styled("li")`
-  /* outline: 1px solid green; */
-`;
+import { Item, ListWrapper, Wrapper } from "./Calendar.styled";
 
 const Calendar = () => {
   const { monthDate, weekDate, isMonthMode } = useCalendar();
