+++ src/components/Slot/Slot.tsx
@@ -1,28 +1,8 @@
+import { DragEvent, FC, useState } from "react";
+
 import { useTrashContext } from "@/hooks/useTrashContext";
 import { TaskType } from "@/types/entities";
-import { DragEvent, FC, useState } from "react";
-import styled, { css } from "styled-components";
-
-export const EmptyItem = styled("li")<{ $hovered: boolean }>`
-  min-height: 10px;
-  flex-shrink: 0;
-
-  ${p =>
-    p.$hovered &&
-    css`
-      min-height: 25px;
-      background-color: rgba(255, 255, 255, 0.138);
-      border-radius: 5px;
-    `}
-
-  transition: min-height 0.3s ease;
-
-  &:last-child {
-    flex-grow: 1;
-    min-height: 30px;
-    border-radius: 20px;
-  }
-`;
+import { EmptyItem } from "./Slot.styled";
 
 interface Iprops {
   index: number;
