+++ src/main.tsx
@@ -1,13 +1,14 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./styles/index.css";
+import { Provider } from "react-redux";
+import { PersistGate } from "redux-persist/integration/react";
+
 import App from "./components/App";
 import GlobalStyle from "./styles/globalStyles";
-import { Provider } from "react-redux";
-import { persistor, store } from "./redux/store";
-import { PersistGate } from "redux-persist/integration/react";
 import HolidaysProvider from "./components/HolidaysProvider";
 import TrashProvider from "./components/TrashProvider";
+
+import { persistor, store } from "./redux/store";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
