+++ src/redux/holidays/holidaysApi.ts
@@ -1,4 +1,5 @@
 import { createApi } from "@reduxjs/toolkit/query/react";
+
 import { axiosBaseQuery } from "@/services/api";
 import { Holiday } from "@/types/entities";
